+++ modules/client/examples/00-client/07-get-transfers.ts
@@ -25,7 +25,9 @@ const params: ITransferQueryParams = {
   direction: SortDirection.ASC, // optional
   type: TransferType.DEPOSIT, // optional
 };
-const transfers: Transfer[] | null = await client.methods.getDaoTransfers(params);
+const transfers: Transfer[] | null = await client.methods.getDaoTransfers(
+  params,
+);
 console.log(transfers);
 
 /*
@@ -44,7 +46,7 @@ console.log(transfers);
     reference: "withdrawing from dao to:0xc8541aAE19C5069482239735AD64FAC3dCc52Ca2",
     transactionId: "0xdb0f9422b5c3199021481c98a655741ca16119ff8a59571854a94a6f31dad7ba",
     to: "0xc8541aae19c5069482239735ad64fac3dcc52ca2",
-    proposalId: "0x1234567890123456789012345678901234567890_0x0"
+    proposalId: "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000"
   },
   {
     type: "deposit",