+++ modules/client/examples/01-tokenVoting-client/06-can-vote.ts
@@ -18,10 +18,10 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 const client = new TokenVotingClient(contextPlugin);
 
 const voteParams: ICanVoteParams = {
-  address: "0x1234567890123456789012345678901234567890",
-  proposalId: "0x1234567890123456789012345678901234567890_0x1",
+  proposalId:
+    "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000",
   pluginAddress: "0x1234567890123456789012345678901234567890",
-  vote: VoteValues.YES
+  vote: VoteValues.YES,
 };
 
 const canVote = await client.methods.canVote(voteParams);