+++ modules/client/examples/01-tokenVoting-client/08-get-proposal.ts
@@ -18,7 +18,8 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 // Create a TokenVoting client
 const client = new TokenVotingClient(contextPlugin);
 
-const proposalId = "0x12345...";
+const proposalId =
+  "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000";
 
 const proposal: TokenVotingProposal | null = await client.methods.getProposal(
   proposalId,
@@ -26,7 +27,7 @@ const proposal: TokenVotingProposal | null = await client.methods.getProposal(
 console.log(proposal);
 /*
 {
-  id: "0x12345...",
+  id: "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000",
   dao: {
     address: "0x1234567890123456789012345678901234567890",
     name: "Cool DAO"