+++ modules/client/examples/06-multisig-client/04-approve-proposal.ts
@@ -18,8 +18,8 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 const client = new MultisigClient(contextPlugin);
 
 const approveParams: ApproveMultisigProposalParams = {
-  proposalId: 0,
-  pluginAddress: "0x1234567890123456789012345678901234567890",
+  proposalId:
+    "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000",
   tryExecution: true,
 };