+++ modules/client/examples/06-multisig-client/06-can-approve.ts
@@ -16,9 +16,9 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 // Create an multisig client
 const client = new MultisigClient(contextPlugin);
 const canApproveParams: CanApproveParams = {
-  pluginAddress: "0x1234567890123456789012345678901234567890",
   addressOrEns: "0x1234567890123456789012345678901234567890",
-  proposalId: 0,
+  proposalId:
+    "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000",
 };
 
 const canApprove = await client.methods.canApprove(canApproveParams);