+++ modules/client/examples/06-multisig-client/09-get-proposal.ts
@@ -16,7 +16,8 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 // Create an multisig client
 const client = new MultisigClient(contextPlugin);
 
-const proposalId = "0x12345...";
+const proposalId =
+  "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000";
 
 const proposal: MultisigProposal | null = await client.methods.getProposal(
   proposalId,
@@ -24,7 +25,7 @@ const proposal: MultisigProposal | null = await client.methods.getProposal(
 console.log(proposal);
 /*
 {
-  id: "0x12345...",
+  id: "0x1234567890123456789012345678901234567890_0x0000000000000000000000000000000000000000000000000000000000000000",
   dao: {
     address: "0x1234567890123456789012345678901234567890",
     name: "Cool DAO"